Moved defs.rs to new config dir.
