+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "seqtoid-pipelines"
-version = "0.1.0"
+version = "0.1.1"
 edition = "2024"
 
 [dependencies]
