+++ src/lib.rs
@@ -1,4 +1,5 @@
 // src/lib.rs
+pub mod config;
 pub mod utils;
 pub mod pipelines;
 pub mod cli;
