+++ src/pipelines/db.rs
@@ -6,7 +6,7 @@ use crate::cli::Arguments;
 use crate::utils::fastx::{read_and_interleave_sequences};
 use crate::utils::file::{file_path_manipulator, extension_remover};
 use crate::utils::command::check_version;
-use crate::utils::defs::H5DUMP_TAG;
+use crate::config::defs::H5DUMP_TAG;
 use crate::utils::db::{write_sequences_to_hdf5, build_new_in_memory_index, check_db};
