+++ src/utils/mod.rs
@@ -3,6 +3,5 @@ pub mod file;
 pub mod streams;
 pub mod command;
 pub mod sequence;
-pub mod defs;
 pub mod db;
